separate into new test suite
