+++ tests/modules/Makefile
@@ -79,6 +79,7 @@ TEST_MODULES = \
     publish.so \
     usercall.so \
     postnotifications.so \
+    postnotifications_perkey.so \
     moduleauthtwo.so \
     rdbloadsave.so \
     crash.so \
